CMake: Audited linkage of 3rdparty dependencies.
This patch forward the `CMAKE_POSITION_INDEPENDENT_CODE` property to
3rdparty projects built with CMake.

It marks as `GLOBAL` imported (non-interface) libraries to enable Cotire
to find the targets. Note that while "global" is usually frowned upon,
it makes sense for these targets as they would be top-level targets if
they were not also 3rdparty targets, causing us to keep them separate.

It adds `IMPORTED_IMPLIB` locations for dependencies which can be built
as shared libraries (`BUILD_SHARED_LIBS=ON`) so that CMake can link to
them correctly. For the libraries that cannot be built in this way, the
linkage was set explicitly and appropriately (e.g. Googletest must
always be built statically).

Review: https://reviews.apache.org/r/61752/