CMake: Bumped minimum required CMake version to 3.7.0.
This is required for the `SOURCE_SUBDIR` functionality of
`ExternalProject_Add`, which is required to avoid a work-around where we
re-implement logic the module can do automatically.

This will also allow us to print timestamps in a cross-platform
manner and append to strings in an obvious manner; among other things.

Review: https://reviews.apache.org/r/61313/