Added a stream ID to resource provider manager connections.
To be able to resolve problems with resource providers that erroneously
try to 'SUBSCRIBE' multiple times a unique stream ID is associated with
a connected resource provider. Resource providers are expected to send
this stream ID with every call that isn't a 'SUBSCRIBE' call.

Review: https://reviews.apache.org/r/61654/