Added IntervalSet to Ranges conversion helpers.
Added a new `common/values.hpp` header file to expose IntervalSet
to Ranges conversion helper declarations.

The most common use of Range resources is for representing network
ports. Since ports are bounded to uint16_t it is awkward to store
them in a IntervalSet<uint64_t>. To address this, convert the
IntervalSet helpers to templates so that we can convert between
IntervalSets of the appropriate type.

Review: https://reviews.apache.org/r/60836/