Enabled `DockerContainerizerProcess::usage` for all platforms.
When originally implemented this code was enabled only for Linux.
However, this code should work anywhere Docker is being used, as it
obtains the information by shelling out to the `docker` executable.

This patch fixes the resources information provided by the `mesos-agent`
on Windows platforms for Docker containers by making the `statistics`
JSON object available.

Note that the `cgroupsStatistics` is Linux specific, and so the
pre-processor guard was not removed but reduced in scope for just it.

Review: https://reviews.apache.org/r/61915/