Checkpoint the agent's decision to recover as a new instance.
After a reboot the agent checkpoints new boot ID after recovering and
agent info after registering with the master. If new agent info is
incompatible with the checkpointed one, the agent will recover as a new
one. But if the agent is restarted before registering it will fail to
start again, because new boot ID will already be checkpointed.

This patch fixes it by checkpointing the agent's decision to recover as
a new one by removing the "latest" meta symlink.

Review: https://reviews.apache.org/r/61661/