Windows: Fixed os::getenv to handle empty string contents.
On Windows, there is normally no difference between setting an
environment variable to emptry string and outright deleting the
environment variable.  However, it is possible to differentiate the
two operations.  This changes the stout helper for `getenv` to
properly check if an environment variable exists or not.

This matters because some environment variables like
LIBPROCESS_METRICS_SNAPSHOT_ENDPOINT_RATE_LIMIT will operate differently
based on whether the variable is empty string or not.