CMake: Enabled building Java artifacts.
This adds an option "ENABLE_JAVA" (default OFF) which will add Java
artifacts, such as Java protobufs and the Mesos Jar, to the CMake
build tree.  These artifact can be built via the `mesos-java` target
and, if enabled, will be built as part of the entire tree.

Review: https://reviews.apache.org/r/61306/