Windows: Changed linking of `IPHlpAPI` library.
While the `#pragma comment(lib...)` method works with MSVC, it is
inconsistent with how we link to the rest of the Windows libraries. So
it was deleted and `IPHlpAPI` was added to stout's Windows dependencies.

Furthermore, the `<iphlpapi.h>` header was removed from the individual
files and placed in `windows.hpp` to ensure it is included in the
correct order with respect to the other Windows headers.

Review: https://reviews.apache.org/r/61597/