Updated CMake version in docker build helper.
Since the minimum version of CMake was bumped to 3.7,
this updates the docker build helper to download CMake 3.8
(stable release that has been out for some time).  The reason
we do not use a package repository for CMake is that none
of them offer a version of CMake beyond 3.5.

Review: https://reviews.apache.org/r/62094