Rescinded offers possibly affected by updates to agent total resources.
When an agent changes its resources, the master should rescind any
offers affected by the change. We already performed the rescind for
updates to the agent's oversubscribed resources; this patch adds offer
rescinding when an update an agent's total resources is processed.

While for updates to an agent's oversubscribed resources we currently
only rescind offers containing revocable resources to e.g., reduce
offer churn, for updates to the total we here currently rescind all
offers for resources on the agent.

As an optimization, this patch adds logic to ignore redundant updates
to agent resources.

Review: https://reviews.apache.org/r/62158