Fixed failure of `DefaultExecutorTest.SigkillExecutor` test on Windows.
Processes terminated by `os::killtree()` exit normally with a status
code on Windows, but killed by a signal on Linux. Prior to this patch,
we expected a process to be always killed by a signal, which caused
test failure on Windows. Since we don't care about the exact value of
a returned status code in this test, this check was removed.

Review: https://reviews.apache.org/r/62279/