Parameterized test for nested container launch.
This introduces a new test class that is parameterized based
on the type of API used to launch (1) the parent container,
(2) the nested container, and (3) the content type (like all
other API tests).  The test is also parameterized based
on the enabled set of isolators.

Because this change overlaps with existing nested container
API tests, those test(s) will be removed.

Review: https://reviews.apache.org/r/63056