Fixed usage of 'ATOMIC_FLAG_INIT'.
The C++ standard guarantees that the macro 'ATOMIC_FLAG_INIT' is
usable for initialization of atomic variables; its use in other
context is unspecified.

libcxx defines 'ATOMIC_FLAG_INIT' as '\{false\}' which when used in an
initializer list yields a clang warning,

    warning: braces around scalar initializer \[-Wbraced-scalar-init\]
        lock(ATOMIC_FLAG_INIT)

This patch prevents such a warning by initializing 'atomic_flag'
member variables in the class declaration instead of in an initializer
list.

Review: https://reviews.apache.org/r/61682/