Added the mark agent gone handler on the master.
This change adds the neccessary logic for handling the mark agent
gone call on the master. Once an agent is marked as gone, it's
not allowed to re-register with the Mesos master. GC'ing the
list of gone agents (it can grow unbounded) would be added in
a separate change.

Review: https://reviews.apache.org/r/62478/