Removed the logic for removing the latest symlink on the agent.
This change removes the logic of removing the latest symlink on
receiving the shutdown message from the Mesos master. This ensures
that agents come back with the same agent ID upon a successful
shutdown similar to the behavior when they come back post a reboot.

Review: https://reviews.apache.org/r/62479/