Windows: Disable some new tests that fetch local URIs.
Mesos currently conflates paths and URIs.  On platforms where the
path separator '\' is not equal to the URI separator '/', tests
that rely on `path` helpers for URIs will naturally fail.

This disables some new tests that fail on Windows for this reason.

The tests here should eventually be fixed along with MESOS-6705.

Review: https://reviews.apache.org/r/62735/