CMake: Updated 3rdparty build rules regarding OpenSSL.
Specifically we now forward all possible hints to `FindOpenSSL` to the
`libevent` and `curl` libraries, as they need to use SSL in the same
configured manner. Furthermore, we explicitly always set `curl` to
enable or disable its use of OpenSSL based on our use, as otherwise it
may detect an OpenSSL installation and cause conflicts (this is the same
as we already do for `libevent`.

We also prefer the multi-threaded version of OpenSSL on Windows.

Review: https://reviews.apache.org/r/62740/