Windows: Added OpenSSL support in libprocess.
This patch adds support for building with OpenSSL on Windows. This
requires an installation of OpenSSL, such as 1.0.2L from Shining Light
Productions. Mesos must be configured like:

    -DENABLE_LIBEVENT=ON -DENABLE_SSL=ON
    -DOPENSSL_ROOT_DIR="C:\OpenSSL-Win64"

The use of `_POSIX_HOST_NAME_MAX` was replaced with `MAXHOSTNAMELEN`,
consistent with the rest of libprocess.

Windows does not define `in_addr_t`, but the type as declared by our
cURL dependency for Windows is `unsigned long`.

OpenSSL on Windows requires the adapter module `openssl/applink.c` to
be compiled as part of the consuming project to deal with Windows
runtime library differences. Not doing so manifests itself as the
"no OPENSSL_Applink" runtime error. The OpenSSL FAQ recommends simply
`#include`-ing it in one of the project's source files, e.g.
libprocess's `ssl/utilities.cpp`.

Review: https://reviews.apache.org/r/62511/