Windows: Disabled RestartSlaveRequireExecutorAuthentication test.
This test became enabled when building with OpenSSL on Windows. The
first part of the test failed because of the use of the Linux command
`cat` to launch an indefinitely running task. This was fixed by using
`more` on Windows. However, the test subsequently fails when recoverying
to due "incompatible agent info", an existing bug: MESOS-7604.

Review: https://reviews.apache.org/r/62512/