Synced the MTU for container eth0 in port mapping isolator.
Currently, the MTU for container eth0 uses a default MTU value (which
is 1500). However, the host eth0's MTU is not guaranteed to be the
default value (e.g., on EC2). And this will breaks path MTU discovery.

This patches forces the MTU of the container's eth0 to be the same as
the host eth0. See more details in MESOS-7130.

Review: https://reviews.apache.org/r/62743/