Propagated the container termination to the agent API.
Updated the agent API so that we can propagate information
from the container termination up to the `WaitNestedContainer`
response. We now propagate resources all the way from the
container limitation to the `WaitNestedContainer` response so
that an executor can know specifically which resource limit
violation caused the container termination.

Review: https://reviews.apache.org/r/62643/