Windows: Added `os::get_job_mem` to stout.
This function sums the working set size for each process in a job object
represented by the PID. The Windows API does not support this directly.
Instead, the list of processes in a job object must be obtained, and
then for each process, the memory usage must be queried.

Review: https://reviews.apache.org/r/63274