Moved the logic of setting '/' as rslave into the launch helper.
This patch prepares for the mount propagation support. In the future,
we can no longer set the root mount as rslave bindly because some
mounts need to be shared. Move this logic to the launch helper is a
necessary step. We will add more logic there to selectively mark some
mount as shared, and the rest as slave.

Review: https://reviews.apache.org/r/63211