Fixed the flaky MasterTest.IgnoreOldAgentReregistration.
This test did not settle the clock before advancing it, which
meant that the call to delay had not yet occurred.

Also updated the test to avoid retries of the registration
message, which can lead the agent to think it is registered
after a detection event, although this shouldn't be an issue
for this test.

Review: https://reviews.apache.org/r/63224