Converted SlaveInfo resources from the registry during master recovery.
The intention is for us to work with resources in the `post-reservation-
refinement` format in the master memory, and store them in the master
registry or agent checkpoint in the `pre` format in order to support
downgrades. In this case, we correctly store the `pre` format resources
in the master registry, but we don't convert them back when we read it
out of the registry. This patch converts the resources back to `post`
when we read from the registry.

This simply addresses a tech-debt. The `authorizeResource` function
for example currently handles both formats because it has to account
for the old format that is stored in the recovered agents.

Review: https://reviews.apache.org/r/63232