Fixed the flaky SlaveTest.HTTPSchedulerSlaveRestart.
This test was trying to prevent re-registration retries, but
it resumed the clock too early! As a result, sometimes a second
re-registration occurs and the master will send an
UpdateFrameworkMessage that over-rules the one we're spoofing
in the test. One example occurs in MESOS-7972, and the agent
as a result replies directly to the scheduler since it sees
the scheduler as having a pid.

Review: https://reviews.apache.org/r/63391/