Migrated `3rdparty/mesos-3rdparty` to `mesos/3rdparty`.
We currently keep some of our 3rdparty bundled dependencies in
an external repository in order to reduce the tarball churn in
the Mesos repository which would bloat the git repository.

The intent is for `github.com/3rdparty/mesos-3rdparty` to be
migrated to `github.com/mesos/3rdparty` along with this patch.

Review: https://reviews.apache.org/r/63479