Removed metrics removal from Master::failoverFramework().
When a framework upgrades from a PID based driver to an HTTP based
driver, the master removes its per-principal metrics. When the same
framework downgrades back to a PID based driver, the master doesn't
reinstate those metrics. This causes a crash when the master receives a
message from the failed over framework and tries to increment its
metrics.

This patch fixes the issue by removing metrics removal from framework
failover handling code. Note that it doesn't handle the case when the
framework's principal change. This situation is being dealt with
separately in MESOS-2842.

Review: https://reviews.apache.org/r/62240/