Updated some isolators' usage of ExecutorInfo.
These four isolators refered to the ExecutorInfo field of the given
ContainerConfig to determine the total resources to isolate.  This
changes these isolators to refer instead to the Resources field of the
given ContainerConfig because the resource amounts should be identical
(enforced by the caller to Containerizer::launch, such as the agent)
and in order to support standalone containers, which specify resources
but no ExecutorInfo.

Review: https://reviews.apache.org/r/63515