Windows: Fixed MESOS-6816 to enable `ExecutorEnvironmentVariables`.
By removing the explicit system environment override code, and instead
setting the system environment as the default in the `CreateProcess`
wrapper, the `SlaveTest.ExecutorEnvironmentVariables` can be enabled.
Note that this also required fixing `os::host_default_path()` and using
it, because the test used `PATH = /bin`, which breaks on Windows.

Review: https://reviews.apache.org/r/63816