Fixed TASK_UNKNOWN status ambiguity.
Currently, TASK_UNKNOWN status covers 2 different cases: task is unknown
and the agent is unknown; task is unknown, but the agent is registered.
This makes it ambiguous to frameworks. In the first case the framework
may elect to wait for some time to see if the agent with the task comes
back. In the second case the task is definitely in a terminal state and
should be reported as TASK_GONE.

Review: https://reviews.apache.org/r/63526/