Added Markdown, bug, and reviewers support to `post-reviews.py`.
These options are appropriately forwarded to `rbt`.

We enable Markdown parsing by default, with `--no-markdown` to disable
it; this is safe as most commit formatting is intended to be Markdown,
and unformatted text is valid markdown. This prevents `rbt` from
escaping e.g. ticks used in commit messages that are then manually
removed from reviews on Review Board.

The `--bugs-closed` flag lets users attach a bug automatically to every
review in the chain they're posting. Similarly `--target-people` will
attach the specified reviewers to every review. This reduces the amount
of work it takes to prepare chains of reviews.

We added help to the argument parser of the script, mostly stolen from
the descriptions in `rbt post --help`.

Review: https://reviews.apache.org/r/62734/