Added ZooKeeper leader resolution to CLI.
This change allows users to use the CLI with a Mesos running in high
availability mode. The `zookeeper` field was already here before this
commit, with an `addresses` array and a `path` field. This change
only adds the backend to actually make it usable.

Interacting with ZooKeeper requires a new dependency, kazoo, that has
been added to the list of requirements for the CLI.

Review: https://reviews.apache.org/r/62938/