Updated offer operation handling to set resource versions.
Resource version UUIDs are used to establish a relationship between
operation and the resource they are operating on. For each agent the
master keeps track of its resource version and the resource version of
the agents' resource providers. As resource versions are required in a
'ApplyOfferOperationMessage', the master has to find the resource
version that's matching the resources an operation should apply to and
set it accordingly.

Review: https://reviews.apache.org/r/63620/