Windows: Fixed `SlaveReregisterTaskExecutorIds` test.
Similar to the check tests, with the `cpu` hard cap enabled, the task in
this test failed with an OOM error because PowerShell requires more
resources than were being allocated. The `defaultTaskResourcesString`
was moved into the `MesosTests` base class so that any test creating a
task can use these default resources, instead of a hard-coded string.
Additionally, the resources allocated by default for an agent were
increased so that multiple tasks can be run simultaneously (as this test
does).

Review: https://reviews.apache.org/r/63779