Updated Docker path for Containerizer::launch interface change.
The change in the return type for Containerizer::launch has less of an
impact on the Docker containerizer, as it (currently) will not return
the "new" enum 'ALREADY_LAUNCHED'.

Note that in changing Containerizer::launch, a private helper of the
Docker containerizer `reapExecutor`s return value was changed.
The return value was originally `Future<bool>` because `reapExecutor`
is the final continuation in the launch path so it needed to match
the return value of Containerizer::launch.  However, `reapExecutor`
never returns `false` (only `true` or a Failure).

Review: https://reviews.apache.org/r/63781