Added quota limit to the master API protos.
This introduces a `limit` in `QuotaInfo` and `QuotaRequest`.

A new `UPDATE_QUOTA` master call is added as a replacement of
`SET_QUOTA` and `REMOVE_QUOTA`, which provides the new semantics
of unset limit meaning "no limit" and unset guarantee meaning
"no guarantee".

The new APIs are marked as experimental for now.

Review: https://reviews.apache.org/r/65334