Fixed `wait()` and `destroy()` in composing containerizer.
Previously, `wait()` and `destroy()` methods of composing containerizer
returned a future that might be set to `READY` state while the internal
state of composing containerizer is not yet cleaned up.

This patch adds a `termination` promise to `Container` struct,
which is used to return a future from `wait()` and `destroy()` methods.
This promise is set to `READY` state iff related container is
completely destroyed.

`_destroy()` callback is subscribed to a future from `wait()`, which
is called on related containerizer, to propagate a value to the
`termination` promise and do the cleanup.

Review: https://reviews.apache.org/r/63887/