Fixed a flaky resource provider test.
An agent could run into a registration timeout and try to register a
second time. In that case the master will acknowledge the registration
twice. This is expected. As a result the agent would send
'UpdateSlaveMessage' twice. The test expected the second
'UpdateSlaveMessage' to be the result of a resource provider
registration, hence the test would fail. This has been resolved by
controlling the clock during registration to avoid registration
timeouts.

Review: https://reviews.apache.org/r/64053/