Made sure all true allocator agent updates trigger allocations.
The allocator method 'updateSlave' can be used to modify either an
agent's capabilities, its total resources, or both. It is expected
that this method triggers an allocation whenever a change occurred.

This patch fixes an issue where if the method as implemented in the
hierarchical allocator was called with new capabilities but the current
total resources, not allocation would have been triggered, even though
the capabilities were updated.

Review: https://reviews.apache.org/r/64001/