Do not kill non-partition-aware tasks upon agent reregistration.
The master used to send a ShutdownFrameworkMessage to the agent
to kill the tasks from non-partition-aware framework but this
misuse of ShutdownFrameworkMessage has led to problematic behavior
in MESOS-7215. This commit fixes the problem by not killing the
tasks for non-partition-aware frameworks when an unreachable agent
re-registers. Instead, the master just recovers these tasks.

Review: https://reviews.apache.org/r/61473/