Added a d_type check in containerizer backend validation.
In the unified containerizer, the overlay fs backend depends
on the underlying filesystem populating the `d_type` field in
`struct dirent`. Raise an error if the user specifies overlayfs
as backend on an unsupported filesystem.  Fallback to other
backends in the default case and raise a warning.

Review: https://reviews.apache.org/r/63652/