Updated the top-level `CSIPluginInfo` proto.
The `type` and `name` fields are introduced to uniquely identify a CSI
service bundle. The purpose of the new proto and these two fields is to
decouple a CSI service bundle from a `ResoruceProviderInfo`, so we may
allow tasks to directly use non-resource volumes (such as NFS volumes).

Review: https://reviews.apache.org/r/63904/