Passed versions when launching tasks.
In this patch we inject resource versions into task launch messages
and add verification in the agent. We require that resource versions
of resource providers whose resources are used in a task have not
changed. With that we can make sure to e.g., not use resources created
in speculated operations.

Review: https://reviews.apache.org/r/64299