Made sure only non-terminal operations use resources.
The function 'addOfferOperation' is a high-level function which can be
used to some extent manage the offer operation life cycle in the
master. In addition to managing the lifetime of offer operations, it
also maintains the state of used resources.

We already made sure to never count speculated operations as using
resources in this function. In addition, the function
'updateOfferOperation' is already capable of updating the used
resources when an offer operation becomes terminal.  To complete this
division of responsibilities we in this patch extend the conditional
treatment in 'removeOfferOperation' to also take into account whether
the operation is terminal when deciding on whether to mutate the used
resources.

Review: https://reviews.apache.org/r/64163/