Removed a stale TODO from before frameworks could modify their roles.
Now that frameworks can modify roles, it's possible for them to have
resources allocated to them for roles that they are no longer
subscribed to. We've already updated the code to handle this, so this
removes the stale TODOs from before this case was possible.

Review: https://reviews.apache.org/r/64237