Improved Windows isolators with `struct Info` abstraction.
Replaced the two maps of `ContainerId -> pid` and `ContainerId -> limit`
with a mapping of `ContainerId -> struct Info { pid, limit }`. This
abstraction correctly ties the the `pid` and cpu/mem `limit` together.
Notably this fixes a subtle bug in `prepare` so that it can no longer be
called multiple times if a resource limit hadn't been set.

Furthermore, this patch imports all the types used in the source files
to enchance code conciseness, and then reformatted with `clang-format`.

Review: https://reviews.apache.org/r/64272