Windows: Documented additional step for running JVM code.
The error message returned when the JVM dependencies cannot be found is
difficult to understand, especially for those unfamiliar with runtime
library loading on Windows. Added this information to the existing
documentation. Also clarified why the `JAVA_JVM_VARIABLE` environment
variable needs to be manually set on Windows.

Note that this and the OpenSSL example probably belong in `windows.md`
instead of `cmake-examples.md`.