Added `curl` executable to Windows build.
Some Mesos health checks use the `curl` executable, which is a required
component on Linux, but also shipped on the system. On Windows,
`curl.exe` is not normally available. However, we already build `curl`
for other components which link to `libcurl`. Part of this build also
builds `curl.exe`, so we renamed the `curl` target to
`libcurl` (explicitly meaning the library), and added `curl` as an
imported executable target on Windows.

Note that this uses `add_custom_command(... POST_BUILD)` to inject a
copy step as a post-build side effect of the `ExternalProject_Add`
target which builds cURL. There is no cleaner way to accomplish this as
CMake does not otherwise have an easy way to copy a file as part of an
_imported_ target at build time. The `add_custom_target` command does
not work because it adds a target that will always be built (not what
we're trying to accomplish here), and the other signature of
`add_custom_command` does not produce a target dependency. While the
`OUTPUT` argument would let us use file dependencies, `curl.exe` isn't a
source file that another target can depend on.

Review: https://reviews.apache.org/r/64102/